WIP: add query to createSampleGroup; some errors
-- removed UI link for adding form values filter if one exists for this dataset
-- correctly adds form values filter w/ options to filter list
-- runs the query only when sample group is to be created

Problems:
1) Currently the Future part of createSampleGroup throws the following error when using a query filter:

Error: key $and must not start with '$'

I think this is because of including '$and' in the options: query: value which is to be written to the DB.
Fix is to JSON.stringify it

2) the Future error, when thrown, results in a different error :
Error: Future resolved more than once

Not sure yet what the fix is. I think it's because the callback on the insert both runs future.throw(err)
and then ALSO future.return(newId).