WIP:bugfix: querybuilder now gets all field values
Each CRFs doc tried to write its value to the key indexed by it's
own _id, instead of the correct placeholder _id that we're using
to represent a CRF name. Just goes to show this placeholder _id
was a bad idea & we need to use the metadata collection as well.